ec_commands: Add new EC_CMD_PD_CONTROL command
This commands makes it possible for the AP to reset the PD chip,
for example during ANX7688 update. Another subcommand prevents
further calls to this command (for security reason, we do not
want the AP to be able to reset the PD chip continuously).

We also keep a chip id parameter in the command, for future
expansion (multiple PD chips).

BRANCH=none
BUG=chrome-os-partner:52433
TEST=ectool pdreset

Change-Id: I7a955dd27b65086c21d195a6504aa7392eb0406d
Signed-off-by: Nicolas Boichat <drinkcat@google.com>