usb_serial: improve installation procedure
There are couple of issues with accessing serial ports over CCD
hardware:

  - sometimes usbserial module is not installed, this causes failures
    to install raiden.ko

  - the softlink to the serial device directory in /dev is changing
    depending on were the physical USB connection is, it makes it
    difficult to script access to the links.

This patch adds a helper script which will be run by udev each time
CCD hardware is detected. The helper script makes sure that necessary
modules are installed and also generates a subdirectory name which is
a zero based index of the currently detected CCD instance. So, the
first one is liked to /dev/google/0 the second one to /dev/google/1,
etc.

An installation script is added to compile the kernel module, create a
valid udev rule file from the template (adds real root directory
path), install the generated rule in /etc/udev/rules.d and force udev
to re-read the rules.

BRANCH=none
BUG=none
TEST=reboot a desktop, plug in SuzyQ, observe /deb/google directory:
  $ find /dev/google/
  /dev/google/
  /dev/google/0
  /dev/google/0/serial
  /dev/google/0/serial/EC
  /dev/google/0/serial/Shell
  /dev/google/0/serial/AP
  $

Change-Id: I374a9869c369410a29f35810d8cfa8b06ec097b4
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>