Update patch set 2
Patch Set 2:

We really need to be tightly coupling pp5000, pp3300, and pmic enable together.  With the USB pd chips being powered off of pp3300 we need to be able to indicate to the pd tasks if they are powered or not. I'm not sure how badly things are intertwined to know how easily a concept of 'usb pd chip powered down' is able to be implemented.

Patch-set: 2