Update patch set 1
Patch Set 1:

> I'm not sure what you meant by "dynamic" inheritance, but we can
 > commit this, once you add HOOK_PRIO_INIT_TCPC.

I mean that these 2 structures, and their function pointers, are filled at runtime in anx7688_driver_init:

  struct tcpm_drv anx7688_tcpm_drv;
  struct usb_mux_driver anx7688_usb_mux_driver;

instead of being filled statically as they are for other drivers.

Don't know enough about EC to know if this patch is ok, or if we should just fill in the structs statically.

Patch-set: 1
Label: Code-Review=0