USB Serial: Add simple install script
The install script builds and installs the raiden module, it also copies
the udev rules file into /etc/udev/rules.d and updates the module alias
and dependency information.

The install script will also retrigger udev to process rules for all
devices that have the Google Vendor ID (0x18d1).  This ensures that
any devices that are connected when the install is run will immediately
be available for use (as opposed to requiring that these devices be
unplugged and replugged before use).

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=None
BUG=None
TEST=Remove udev rules file
     Remove raiden.ko module, aliases, and dependency information
     Reboot workstation
     Run ./install
     Plug in CCD capable device
     ls /dev/google

Change-Id: I7bcb02f05ee84738a6259800afc4d0c69bea9e69