pd: Remove PD_STATE_SUSPENDED, and replace by variable
PD_STATE_SUSPENDED does not appear to be used currently, and
is broken if pd_set_suspend is called while the task is running
(as the state might get updated after it is set to suspend).

We replace the state with a boolean in the pd port structure, so
that we can suspend the port properly.

BRANCH=none
BUG=chrome-os-partner:52433
TEST=On elm:
   ectool pdcontrol reset
   sleep 0.1
   ectool pdcontrol suspend
   i2cset -f -y 2 0x28 0x05 0x10
   => No I2C communication with ANX7688

Change-Id: Iad7b500f1af554e2bc4f64f90ee0a492f903749a
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>