Addresses a timer scheduling race condition
This was an issue on Cr50 because the interrupt priority grouping scheme
(see Application Interrupt and Reset Control Register) is currently
configured to treat the numerical priorities as defined in the IRQ
declarations as priority bits (i.e. higher priority interrupts can
preempt lower priority ones). This might or might not be an issue on
other targets.

See crosreview.com/347465 for details.

BRANCH=none
BUG=none
TEST=none

Change-Id: Ic4222ad2615b26fd70e4dedf84672c6cc1ef06de
Signed-off-by: Nadim Taha <ntaha@chromium.org>