motion: Add ability to stop ring interrupts.
Currently, it is assumed the host will sooner or later retrieve the
events from the sensor ring: It is only used by Android and the sensor
HAL is enabling the ring buffer at boot.
But if nobody process the ring, but the ring is almost full, the EC will
generate interrupt for every new events.
This can happen with ARC, where events generated for chrome os
will be in the ring but nobody will process them until Android is
started.

Add a command to allow sending ring MKBP events. It will be used when
the IIO ring buffer is enabled / disabled.
It also can be used for preventing raising interrupt when the device is
about to go to sleep.

BRANCH=ryu,cyan
BUG=b:25425420,b:27849483
TEST=Check with fiforead that no events are queued when IIO ring
buffer is disabled.
Check with ectool and androsensor that interrupt generation stops.
CQ-DEPEND=I9cbda167796e8f70d8092b692ed86c89b3ac2c6d

Change-Id: I1390e40ee977060091f1bc0fce0fb353a9367f90
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>