Update patch set 2
Patch Set 2:

This effectively turns I2C from interrupt-driven to fast-polled unless there is clock-stretching.  I'm not sure that's what we want.

Perhaps it would be better to decide on the behavior based on the I2C bus speed.

If 100 kHz, we can safely sleep for 100 us between checks, as we do now.

If 400 kHz or faster, bytes should go out every ~25 us.  That's too short to use usleep().  So, fast-poll for the first ~50 us and then sleep for 100 us between checks.  If there is clock stretching, we'll spend more of it asleep.

Patch-set: 2
Label: Code-Review=0