Update patch set 2
Patch Set 2:

Yes, I known there is a driver issue for touchscreen. I am afraid there is no time for this project to wait for the driver update. so I think the solution of delaying more time should be a temporary solution. Once the root cause is found and driver updated, we can remove this temporary solution.

Patch-set: 2