charge_manager: Seed charge manager after charger task is initialized
If the battery is connected, PD is updated as unknown initially and
then dual role capable. Hence PD tries to re-negotiate or try source.
Before PD does a re-negotiation, if the charger task is not initialized
and doesn't knows about the battery status, system loses the power.
To avoid this issue, do not seed the charge manager or try source from
PD until the charger task is initialized and knows about the battery
status.

BUG=chrome-os-partner:54058
BRANCH=none
TEST=Amenia system does not reboot before booting to OS.

Change-Id: If8fd84f82f5a7fb7ca3736031a161d90e5e77c12
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>