Update patch set 2
Patch Set 2:

Looks like there is a problem with the additional tasks requiring too much .bss for their stacks. Patch set 3 has reduced the stack sizes to verify this hypothesis. Not intended to be a solution, but need it as part of the CL so I can kick of the nyan_freon trybuild

Patch-set: 2