CR50: add support for writing to info bank
State that needs to survive re-flashing
of RO+RW is stored in the INFO bank.  An
example of such state is manufacture secrets,
which need to survive reflashing from the
personalize firmware to the initial TPM2
firmware.

This change adds support for writing to
the writeable flash info bank.

BRANCH=none
BUG=chrome-os-partner:43025
TEST=manually verified info1 reads and writes

Signed-off-by: nagendra modadugu <ngm@google.com>

Change-Id: I9226e2161e036d1dacccbe55b67724b449983008