Update patch set 2
Patch Set 2:

The commit message explains most of the context. Please pay close attention to the new mutex function since I had to implement it for 4 different cores. It's very similar to the existing mutex lock/unlock functions, but adds a member to track the task number which holds the lock. This way the same task can attempt to grab the lock again without causing the task to suspend itself. 

The motivation for this change is to support the TPM2.0 NvMemory module which is using the scratch memory buffer and needs to protect it but doesn't have a single entry point.

Patch-set: 2