Update patch set 1
Patch Set 1:

Why? To make it simpler for users. They are asked to add a udev rule anyway so extending it to cover the IDs issue seemed reasonable. What race condition do you see?
I don't think a kernel driver for this simple task is appropriate, and it gets out of sync continuously.
And yes, the IDs can be added and they indeed are, but for newer kernel versions than what we have in our desktops currently.

Patch-set: 1