Update patch set 2
Patch Set 2: Code-Review-1

I think this will lead to unintended side effects. For example, in powerbtn_x86_lid_change() -- if we sysjump, this hook will trigger, and we'll power-on the AP if it's off.

In the past, when actions had to be taken based on lid state, we added a HOOK_INIT that checks the lid state directly.

Patch-set: 2
Label: Code-Review=-1