Update patch set 1
Patch Set 1: Code-Review-2

This should not be needed with the new install script, which installs both the udev rule and the raiden module.

More detail: While Vincent is right that you do not strictly need the raiden module, there are in fact many ways to make this work.  One way is as you're doing here and Vincent prefers, which is to add the USB VID/PID pair to the usb-serial module's list of devices.  The problem I have with that is that it will certainly not scale as new devices are added.  That is why I made the CCD serial console forwarding endpoints all use a common subclass/protocol.  This subclass/protocol pair is defined for all Google USB devices, so the only part of the VID/PID that needs to be checked is the VID.  That's what the existing rules in this file do.  Adding a new rule for each new device is busywork that isn't needed.  The raiden module (which is only about five lines of actual code) just adds a matching rule to the usb-serial driver for the Google VID and the CCD serial subclass/protocol.  That means that the raiden module will find any CCD serial device, independent of PID.  This has a number of advantages, new devices will just work with no change, and the SPI forwarding bridge that CCD provides will NOT be mistaken for another serial console (which the VID/PID mechanism used here fails to differentiate).  Also, the raiden module matching rule has been added upstream to the generic usb serial driver.  That means that eventually you can just stop installing the raiden module and things continue to work with the relatively simple udev rule as it stands.  Unless you've added the VID/PID pair to use-serial, then I have no idea who wins, but probably the VID/PID pair will at least continue to incorrectly configure the SPI bridge.

Patch-set: 1
Reviewer: Gerrit User 1112784 <1112784@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2