Update patch set 1
Patch Set 1:

Thanks for the detailed explanation, understood (well, not totally to be honest but mostly).
However, the problem still remains. In old kernels, the raiden module or other solution is needed, but the module is built for specific kernel version and eventually gets out of sync. Some people here even faced not being able to 'make' it (and not everyone knows about kernel stuff). That certainly does not scale either.
Eventually a temporary solution like this will not be needed, so maybe scalability is not really an issue. But I am all for finding whatever temporary solution, accepting some tradeoffs, as long as it is practical for the user.

Patch-set: 1