Update patch set 1
Patch Set 1:

> Perhaps the solution is to add Anton's explanation to a README, and
 > include Julio's workaround as example code within that README.

Yes, and we can do one better, by having the install script select one of the two solutions based on whether the module should be used.  I was not aware that there were failures to build the module.  I assume that that is because of really old kernels?  In which case we should be able to figure out which kernel version made the changes required and fall back for older kernels.  It might also be that the kernel headers are not installed, which could also trigger a fallback.  I would recommend that the fallback path be a separate udev rule that uses the same VID/subclass/protocol matching mechanism to add the device VID/PID pair so that it doesn't have to be updated for each new device (the PID would be extracted from the sysfs database that udev maintains).

Also, it's clear that I drastically under-documented my rational for these decisions, I'll add a README as Bill suggests.  And I'll look at writing the fallback udev rule file.  But I'll need your help (Julio) to identify when that path should be used.

Thanks all
    Anton

Patch-set: 1