USB Serial: Add fallback rule and script
This rule makes it easier to use CCD devices when the raiden module
can't be installed for some reason.  The rule informs the usbserial
module that it should handle anythink that looks like a simple serial
port for any CCD compatible USB devices.

The install script now detects failure when building and installing
the raiden module and offers the --fallback option in that case.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=None
BUG=None
TEST=make buildall -j

Change-Id: I617bbdfb4c5cb9e9803f4088c651f84e3f72bd28