Update patch set 2
Patch Set 2: Code-Review-1

It's fine to have a debug_printf() for your routines, but it should be a wrapper around cprintf() or cprints() - or better still, a varargs macro like CPRINTF().  And use an appropriate output channel (or define one, carefully).

We did this because otherwise one task with a lot of debug output could overwhelm the console and prevent debugging other tasks, and it was a pain when debugging one task to have to comment out all the printf()'s from unrelated tasks.

Patch-set: 2
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1