Update patch set 2
Patch Set 2:

> It's fine to have a debug_printf() for your routines, but it should
 > be a wrapper around cprintf() or cprints() - or better still, a
 > varargs macro like CPRINTF().  And use an appropriate output
 > channel (or define one, carefully).
 > 
 > We did this because otherwise one task with a lot of debug output
 > could overwhelm the console and prevent debugging other tasks, and
 > it was a pain when debugging one task to have to comment out all
 > the printf()'s from unrelated tasks.


This is my reading of the code: cprintf is defined in common/console_output.c, and this file is conditionally compiled when the CONSOLE task has been enabled.  Is this correct?  The change below applies to the RO section, which currently does not enable the CONSOLE task.  Should I be enabling the CONSOLE task in RO?

(console_output.c includes "DECLARE_CONSOLE_COMMAND()" directive)

Patch-set: 2