Update patch set 1
Patch Set 1:

OK, I've made some significant changes.  This version only adds the VID:PID device ID pair if it isn't already in the device ID list for usbserial.  To do that I had to make a helper script, and that needed to be installed, so I've updated the install script to have a --fallback option that will do that (and install the udev rule to trigger the script).  I've also added error checking for the building and installing of the raiden module so the user can be informed that they might want to use the --fallback option (which I opted to not do for them by default if the module path fails).  There are enough corner cases with this now that I did a bunch of testing, but probably missed a case or two.  Please take a look and test it if you can.  In particular, Julio, if you have a module building failure case I'd love you to try the install script there.  I forced a build failure and it did what I expected, but catching the in field failure would be better.

Patch-set: 1