Update patch set 1
Patch Set 1: Verified-1

Just noting that this doesn't work right now.  If nothing else it reboots your AP at resume, but even after that is fixed something is still unhappy in the kernel.

It's probable that this unhappiness is due to a kernel problem (the kernel may not be saving / reinitting certain things yet).

Patch-set: 1
Label: Verified=-1