Update patch set 2
Patch Set 2:

The module path works for me. I will find someone tomorrow to test fallback path.
I think an issue still remains. The module path (installed module) will not work if/when a new kernel is installed, and the user might not understand why CCD stopped working. Hopefully they will run the install script again.
If I am not wrong on that, then why not use only fallback method? I agree the module is the way to go if it is included in the kernel, but here we are looking for a solution when it is not included, and the fallback method starts to look good enough.

Patch-set: 2
Reviewer: Gerrit User 1129901 <1129901@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0