Update patch set 2
Patch Set 2:

(1 comment)

Julio: I personally prefer the module method over the fallback method, even if it means having to reinstall the module when the kernel is bumped.  But I can see how that could be confusing for an end user.

Vadim: Sure, doing the line escaping at install time could be done.  I'm not sure it would be any less confusing though, because it would mean that there is yet another format that someone familiar with writing udev rules would have to grok.

All: But, since neither of these ideas directly effect the functionality of this CL, and since today is my last day on Chrome OS.  :)  Could we please land this and then make the proposed changes (if you deem them required) in later CLs?

Thanks,
    Anton

Patch-set: 2
Label: Code-Review=0