Cave: Enable Gyro
The range of odr is from 25 Hz to 3200 Hz.
The odr value is set as 1600 Hz to fix gyro function fail.

Changing the array order is due to the shared memory addresses
are predefined to the fixed addresses. Without rearranging the
array order, the kernel sensor driver will read the wrong data.

BUG=chrome-os-partner:53714
BRANCH=firmware-glados-7820.B
TEST=emerge ec, and gyro function is OK.

Change-Id: I6b5388502bf1d13a9efd0978ad93a2546e0b45e0
Signed-off-by: Shasha Zhao <Sarah_Zhao@asus.com>
Signed-off-by: Zhuo-hao Lee <zhuo-hao.lee@intel.com>