Update patch set 1
Patch Set 1:

(2 comments)

> I'm not sure board_pd_idle() provides all the semantics we need? Essentially one needs to wait for the pd state machine to indicate idle. Then the board has a chance to suspend the port. However, there's no way to introduce policy on forcing such a situation... at least I'm not sure how it's currently done.

Yes, when relying on board_pd_idle(), we must wait for the port to become idle. I think this fits the reef requirements.

> I'm also a little concerned about the callback approach because we currently have multiple tasks handling each pd port. As such things can get racey. Have you given much thought of how we'd handle reef w/ 2 ports plus power domains which are interdependent?

Yes, see my previous comment:

You want to take action only when both ports are idle. You will get the CB called every 1 second when a port is idle + not suspended, so it's up to the board level to determine that both ports are idle -- and since we have no "not idle" callback, you might need to check timestamps. If this is too hacky, we can add a "not idle" callback / parameter, but I think this is going to be racy.


Anyway, pd_set_suspend() has now been changed to make 3 attempts at setting the PD state machine to suspended state, so we might consider calling it outside of the PD tasks now.

Patch-set: 1
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0