Update patch set 1
Patch Set 1: Code-Review+2

It should, but adding a 50K in parallel with the existing 4.7K pulldown would drop the resistance to 4.3K. Check with Koon to be sure that's okay, otherwise we need to distinguish between them.

And we do plan for Gru to have the correct pulldowns added in the next build, right? If so, file a bug to undo this when that happens.

Patch-set: 1