driver: bmi160: Eliminate port-specific #defines
This patch replaces CONFIG_SPI_ACCEL_PORT and I2C_PORT_ACCEL with
CONFIG_SPI and CONFIG_I2C. In the places where they are used
BMI160_IS_SPI(addr) decides the code path. The port is passed in
as a function arg, so having it in a #define is redundant.

This also makes the code more suitable for devices with multiple
sensors, for example, an accelerometer in the lid and also in the
base.

BUG=none
BRANCH=none
TEST=built for Reef without I2C_PORT_ACCEL defined, BMI160 initialized
successfully.

Change-Id: Ifc4acb4776e661172b840e535a9f7e4fbbdd004e
Signed-off-by: David Hendricks <dhendrix@chromium.org>