cr50: add RO image hash base to the reported FW version
The CR50 RO version is identified not just by the git hash, but also
by the fuse settings and keys used for signing. The first four bytes
of the entire RO image's hash are saved in the image header. Adding
these four bytes to the version string reported to the host allows to
uniquely identify both RO and RW firmware versions.

BRANCH=none
BUG=none
TEST=verified that the appropriate string is showing up:
  localhost ~ # grep cr50 /sys/firmware/log
  Firmware version: RO: 97594095 RW: cr50_v1.1.4803-dcac93a-dirty
  localhost ~ #

Change-Id: I30a21fad15d99523b1edfa1baa32d80b44e7d0df
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>