Update patch set 1
Patch Set 1:

You can only write all 0s to memory if it has not been erased. I used a file with all 0s so I did not run into that error. I think that is ok. I will just add that as a requirement in the Lucid firmware updater.

Patch-set: 1