reef: Drive PMU_RSTBTN_N manually
Due to what appaers to be a leakage issue, this patch drives the
SYS_RST_ODL (aka PMU_RSTBTN_N) pin low until after PCH_PWROK gets
asserted. After that, we drive the signal high and hold it for a
while, then re-configure the pin as open-drain to let it float.

BUG=chrome-os-partner:53791
BRANCH=none
TEST=SoC boots when PWROK is passed in much earlier than it used to be

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I3429263f1a115e0a745a7014089ab0682e25d7cf