g: disable uart 1 and 2 during init
When the uart rx signal is not externally pulled high by the EC or AP, 
the low rx signal triggers thousands of uart interrupts. At 
initialization Cr50 does not know the state of those devices. If the 
uart is initialized when the device is off these interrupts may prevent 
Cr50 from booting on certain boards. This change does not enable the 
uart until the device state is know. When the device state monitoring 
detects that the AP or EC is powered on it will enable uart 1
or 2 and when it detects that it is powered off then the uart will be 
disabled.

BUG=none
BRANCH=none
TEST=UART_CTRL registers are set to 0 for uart 1 and 2, and are changed
to 3 when the device state is on.

Change-Id: I43e847c6abb8507a86de92a5c226a79f3add7f97
Signed-off-by: Mary Ruthven <mruthven@chromium.org>