Update patch set 2
Patch Set 2: Verified+1

Yep, setting the pin as GPIO_ODR_LOW works as well, so I went ahead and made that change.

Unfortunately just setting it to high and letting it float wasn't having the effect we had hoped for. Something seems to be pulling it down such that the external pull-up doesn't pull it up as fast as we'd like. That's why I set the pin as push-pull and drive it high for a while. It's a bit of a mystery as to why this is happening, but it seems to have some effect on our ability to bring the SoC up.

Driving the pin high (push-pull) for about 1sec and then letting it float (open-drain) seems to do the trick, though.

Patch-set: 2
Reviewer: Gerrit User 1000253 <1000253@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1