Update patch set 11
Patch Set 11:

(1 comment)

Still not quite understand the reason to remove the alternate definition from gpio.inc and add code in board.c to switch to i2c mode.

When EC reset, it will most likely auto boot to S0 unless the battery is not enough charge or have charger issue. In the case that is blocked to boot, als and motionsense will not access i2c, i2cscan will directly return busy, i2cxfer will unwedge and fail but either the code that access i2c should handle the error or the user run from EC console should understand the error, so I don't see the need to not do the alternate definition in gpio.inc.

Patch-set: 11