reef: Rail and PMIC init changes for newer boards
Rev1 brought up 5V, 3.3V, and PMIC very early in the EC boot process
due to dependencies in the power topology. This had some other side-
effects, for example, a lot of the power rails would already be up by
the time the EC got around to processing the power state machine thus
thus leaving it waiting for signal changes that were supposed to come
later but had already occurred instead.

This patch adds an alternate path for newer boards so that the power
sequencing is matches what we really expect.

BUG=chrome-os-partner:54962
BRANCH=none
TEST=built and booted on multiple revisions of reef (including rev1)

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: If9ddd41044f132e719b0b7f0ab80ed908ddb1d9b