mkbp: Add support for reporting non-matrixed keys.
Currently, the matrix keyboard protocol does not have support for
handling non-matrixed keys.  This commit adds support for buttons which
do not appear in the keyboard matrix.  The events will still share the
same keyboard FIFO, however some of the events may simply be the new
state of the buttons on the system instead of the key matrix.

BUG=chrome-os-partner:54988
BUG=chrome-os-partner:54976
BRANCH=None
TEST=Flash kevin, and verify that keyboard is still functional.
TEST=make -j buildall

Change-Id: If4ada904cbd5d77823a0710d4671484b198c9d91
Signed-off-by: Aseda Aboagye <aaboagye@google.com>