Update patch set 1
Patch Set 1:

(1 comment)

Are you planning on landing this? Or is this one we'll cherry-pick for a local build specifically for EVT? From what I can see of the changes we'll need to cherry-pick CLs locally and do EC builds.

Please mark this -2 yourself. Reviewers will +2 and then we can update the bugs w/ the CLs that address the EVT changes. And then we'll have a recipe for people doing a local build.

Patch-set: 1
Label: Trybot-Ready=0