CR50: add logic for handling ack / cert injection at manufacture
This change adds a logic for handling injected messages
during stage two of manufacture.  In this stage, there
are two steps.

1) an "ack" command is sent to the chip, to which the chip
   responds with a message corresponding to its ID.
2) in the second step, the chip is sent a message containing
   its certificates.  These are verified for correctness,
   and stored in NVRAM.

The chip then marks itself as manufactured, following which
TPM commands may be issued to the chip.

Manufacture related commands are non-standard, and are
implemented under the EXTENSION command set.

Also filed crosbug/p/55260, in which the TPMT_PUBLIC
template is truncated during personalization,
and hence requires a workaround in RW firmware.

Also, some changes to logic that tests for the current
state of manufacture:

The tpm_manufactured() check verifies if the proper certificates are
found in the device NV RAM.  This check can not succeed unless NV RAM
metadata is initialized by calling _TPM_Init().

If tpm_manufactured() shows that the device has not been through
the manufacturing sequence yet, TPM_Manufacture() needs to be invoked to
make sure that all relevant TPM structures are initialized and properly
stored in NVRAM. _TPM_Init() needs to be invoked again after that.

BRANCH=none

BUG=chrome-os-partner:43025,chrome-os-partner:47524
BUG=chrome-os-partner:50115,chrome-os-partner:55260
TEST=build succeeds

Change-Id: I63f87af30dff07afe39549015274cbe4fff20049
Signed-off-by: nagendra modadugu <ngm@google.com>