CR50: add logic for handling ack / cert injection at manufacture
This change adds a logic for handling injected messages
during stage two of manufacture.  In this stage, there
are two steps.

1) an "ack" command is sent to the chip, to which the chip
   responds with a message corresponding to its ID.
2) in the second step, the chip is sent a message containing
   its certificates.  These are verified for correctness,
   and stored in NVRAM.

The chip then marks itself as manufactured, following which
TPM commands may be issued to the chip.

Manufacture related commands are non-standard, and are
implemented under the EXTENSION command set.

BRANCH=none

BUG=chrome-os-partner:43025,chrome-os-partner:47524,chrome-os-partner:50115
TEST=build succeeds

Change-Id: I8b41449fc2be83291f791e353003d9c32ce64573
Signed-off-by: nagendra modadugu <ngm@google.com>