cr50: do a soft reset instead of hard when responding to sys_rst_l
Soft reset will not power down Cr50, so it wont assert EC_RST_L when it
reboots. This is a short term solution that fixes the cold vs warm reset
issue, but it does not fix the problem with all of all of the other Cr50
functions disappearing during warm reset. We will need to just reset the
TPM functionality on Cr50 to fix those issues.

BUG=chrome-os-partner:52366
BRANCH=none
TEST=Asserting SYS_RST_L resets Cr50 and the AP, but the EC doesn't
reset and Cr50 TPM functionality still works.

Change-Id: I7c0d0712ee6df59cfadd3975caf3503d44f3eff2
Signed-off-by: Mary Ruthven <mruthven@chromium.org>