Update patch set 9
Patch Set 9:

This patch is difficult to live with in real life environment - we need a functional tpm to boot into chrome os berore manufacturing function can be run.

How about we do not prevent TPM from functioning if it is not manufactured, but we reinitialize it on every startup. And once it has been through manufacture we reboot, and then just reject the following manufacture request.

Patch-set: 9
Label: Code-Review=0